# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades/{tradeId}/executions:
    post:
      operationId: createExecution
      tags: [executions]
      summary: Add an execution to a trade
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionInput"
      responses:
        "201":
          description: Created — returns full trade with recalculated values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
        "400":
          description: Invalid execution (e.g. exits exceed open position)
        "404":
          description: Trade not found
  /executions/{id}:
    patch:
      operationId: updateExecution
      tags: [executions]
      summary: Update an execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionUpdate"
      responses:
        "200":
          description: Updated — returns full trade with recalculated values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteExecution
      tags: [executions]
      summary: Delete an execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted — returns full trade with recalculated values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"
        "404":
          description: Not found

  # ─── TRADE REVIEW ────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Headline performance statistics
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Stats grouped by currency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DashboardStats"
  /dashboard/equity-curve:
    get:
      operationId: getDashboardEquityCurve
      tags: [dashboard]
      summary: Cumulative P&L over time
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Equity curve data points grouped by currency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquityCurveGroup"
  /dashboard/pnl-by-day:
    get:
      operationId: getDashboardPnlByDay
      tags: [dashboard]
      summary: P&L grouped by day of week
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: P&L by day of week grouped by currency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PnlByDayGroup"
  /dashboard/pnl-by-strategy:
    get:
      operationId: getDashboardPnlByStrategy
      tags: [dashboard]
      summary: Performance metrics grouped by strategy
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Strategy performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyPerformance"
  /dashboard/monthly-pnl:
    get:
      operationId: getDashboardMonthlyPnl
      tags: [dashboard]
      summary: P&L grouped by month
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly P&L grouped by currency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyPnlGroup"
  /dashboard/recent-trades:
    get:
      operationId: getDashboardRecentTrades
      tags: [dashboard]
      summary: Most recent closed trades
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeSummary"
  /dashboard/win-loss:
    get:
      operationId: getDashboardWinLoss
      tags: [dashboard]
      summary: Win/loss/breakeven breakdown
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Win/loss breakdown grouped by currency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WinLossGroup"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Confluence:
      type: object
      required: [id, categoryId, categoryName, name, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        categoryName: { type: string }
        name: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Tag:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        createdAt: { type: string, format: date-time }
...
    Execution:
      type: object
      required: [id, tradeId, side, executedAt, price, quantity, quantityUnit, createdAt, updatedAt]
      properties:
        id: { type: integer }
        tradeId: { type: integer }
        side: { type: string, enum: [buy, sell] }
        executedAt: { type: string, format: date-time }
        price: { type: number }
        quantity: { type: number }
        quantityUnit: { type: string }
        commission: { type: ["number", "null"] }
        fees: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ExecutionInput:
      type: object
      required: [side, executedAt, price, quantity, quantityUnit]
      properties:
        side: { type: string, enum: [buy, sell] }
        executedAt: { type: string, format: date-time }
        price: { type: number }
        quantity: { type: number }
        quantityUnit: { type: string }
        commission: { type: number }
        fees: { type: number }
        notes: { type: string }
    ExecutionUpdate:
      type: object
      properties:
        side: { type: string, enum: [buy, sell] }
        executedAt: { type: string, format: date-time }
        price: { type: number }
        quantity: { type: number }
        quantityUnit: { type: string }
        commission: { type: ["number", "null"] }
        fees: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }

    # ── TRADE REVIEW ─────────────────────────────────────────────────────────
    TradeReview:
      type: object
      required: [id, tradeId, createdAt, updatedAt]
      properties:
        id: { type: integer }
        tradeId: { type: integer }
        reviewNotes: { type: ["string", "null"] }
        whatWentWell: { type: ["string", "null"] }
        whatWentWrong: { type: ["string", "null"] }
        ruleAdherence: { type: ["integer", "null"] }
        tradeQuality: { type: ["integer", "null"] }
        lessonsLearned: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TradeCalculated:
      type: object
      description: Values derived from executions — never stored, always computed
      required: [avgEntry, openPositionSize, realizedGrossPnl, realizedNetPnl, totalCommissions, totalFees, tradeLevelCosts, openedAt]
      properties:
        avgEntry: { type: ["number", "null"] }
        avgExit: { type: ["number", "null"] }
        openPositionSize: { type: number }
        realizedQuantity: { type: number }
        realizedGrossPnl: { type: number }
        realizedNetPnl: { type: number }
        totalCommissions: { type: number }
        totalFees: { type: number }
        tradeLevelCosts: { type: number }
        plannedR: { type: ["number", "null"] }
        actualR: { type: ["number", "null"] }
        openedAt: { type: ["string", "null"], format: date-time }
        closedAt: { type: ["string", "null"], format: date-time }
        durationMinutes: { type: ["integer", "null"] }
        outcome: { type: ["string", "null"], enum: [win, loss, breakeven, null] }

    # ── TRADE SUMMARY (for journal list) ─────────────────────────────────────
    TradeSummary:
      type: object
      required: [id, accountId, accountName, instrumentId, instrumentSymbol, assetClass, settlementCurrency, direction, status, tradeTimezone, createdAt, updatedAt]
      properties:
        id: { type: integer }
        accountId: { type: integer }
        accountName: { type: string }
        instrumentId: { type: integer }
        instrumentSymbol: { type: string }
        instrumentName: { type: string }
        assetClass: { type: string }
        settlementCurrency: { type: string }
        direction: { type: string, enum: [long, short] }
        status: { type: string, enum: [open, partially_closed, closed] }
        strategyId: { type: ["integer", "null"] }
        strategyName: { type: ["string", "null"] }
        setupId: { type: ["integer", "null"] }
        setupName: { type: ["string", "null"] }
        tradeTimezone: { type: string }
        notes: { type: ["string", "null"] }
        tradeLevelCostAdjustment: { type: ["number", "null"] }
        plannedEntry: { type: ["number", "null"] }
        initialStopLoss: { type: ["number", "null"] }
        profitTarget: { type: ["number", "null"] }
        plannedPositionSize: { type: ["number", "null"] }
        plannedRisk: { type: ["number", "null"] }
        plannedRiskPercentage: { type: ["number", "null"] }
        calculated:
          $ref: "#/components/schemas/TradeCalculated"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    # ── TRADE DETAIL (full, with executions/confluences/tags/review) ──────────
    TradeDetail:
      allOf:
        - $ref: "#/components/schemas/TradeSummary"
        - type: object
          required: [executions, confluences, tags]
          properties:
            executions:
              type: array
              items:
                $ref: "#/components/schemas/Execution"
            confluences:
              type: array
              items:
                $ref: "#/components/schemas/Confluence"
            tags:
              type: array
              items:
                $ref: "#/components/schemas/Tag"
            review:
              $ref: "#/components/schemas/TradeReview"
...
    DashboardStats:
      type: object
      required: [currency, tradeCount, winCount, lossCount, breakevenCount, winRate, netPnl, grossPnl, avgWinner, avgLoser, avgR, profitFactor, expectancy]
      properties:
        currency: { type: string }
        tradeCount: { type: integer }
        winCount: { type: integer }
        lossCount: { type: integer }
        breakevenCount: { type: integer }
        winRate: { type: number }
        netPnl: { type: number }
        grossPnl: { type: number }
        avgWinner: { type: number }
        avgLoser: { type: number }
        avgR: { type: ["number", "null"] }
        profitFactor: { type: ["number", "null"] }
        expectancy: { type: number }
    EquityCurvePoint:
      type: object
      required: [date, cumulativePnl, tradeId]
      properties:
        date: { type: string, format: date-time }
        cumulativePnl: { type: number }
        tradePnl: { type: number }
        tradeId: { type: integer }
    EquityCurveGroup:
      type: object
      required: [currency, points]
      properties:
        currency: { type: string }
        points:
          type: array
          items:
            $ref: "#/components/schemas/EquityCurvePoint"
    PnlByDayEntry:
      type: object
      required: [dayOfWeek, dayName, netPnl, tradeCount, winRate]
      properties:
        dayOfWeek: { type: integer }
        dayName: { type: string }
        netPnl: { type: number }
        tradeCount: { type: integer }
        winRate: { type: number }
    PnlByDayGroup:
      type: object
      required: [currency, days]
      properties:
        currency: { type: string }
        days:
          type: array
          items:
            $ref: "#/components/schemas/PnlByDayEntry"
    StrategyPerformance:
      type: object
      required: [strategyId, strategyName, currency, tradeCount, winRate, netPnl, avgR]
      properties:
        strategyId: { type: ["integer", "null"] }
        strategyName: { type: string }
        currency: { type: string }
        tradeCount: { type: integer }
        winRate: { type: number }
        netPnl: { type: number }
        avgR: { type: ["number", "null"] }
    MonthlyPnlEntry:
      type: object
      required: [month, year, netPnl, tradeCount]
      properties:
        month: { type: integer }
        year: { type: integer }
        netPnl: { type: number }
        tradeCount: { type: integer }
    MonthlyPnlGroup:
      type: object
      required: [currency, months]
      properties:
        currency: { type: string }
        months:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyPnlEntry"
    WinLossEntry:
      type: object
      required: [outcome, count, totalPnl]
      properties:
        outcome: { type: string, enum: [win, loss, breakeven] }
        count: { type: integer }
        totalPnl: { type: number }
    WinLossGroup:
      type: object
      required: [currency, breakdown]
      properties:
        currency: { type: string }
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/WinLossEntry"

    # ── SETTINGS ─────────────────────────────────────────────────────────────
...